Move PolicyPage.jsx contents into index.jsx
